Revert "UI Fixes"
This reverts commit cd261a9d283f684188361a1049c1e9f43d97af7f.

+++ ArcDice/Assets/HUDManager.cs
@@ -77,7 +77,7 @@ public class HUDManager : MonoBehaviour
 
     public void UpdateHealth()
     {
-        healthText.text = health + (int)playerManager.Health;
+        healthText.text = health + playerManager.Health;
     }
 
     private void Update()
